NioSqlResourceManagerImpl rename to SqlResourceManagerImpl and old SqlResourceManagerImpl and SqlLoader removed.
