+++ src/main/java/jp/co/future/uroborosql/UroboroSQL.java
@@ -34,8 +34,8 @@ import jp.co.future.uroborosql.filter.SqlFilterManager;
 import jp.co.future.uroborosql.filter.SqlFilterManagerImpl;
 import jp.co.future.uroborosql.mapping.DefaultEntityHandler;
 import jp.co.future.uroborosql.mapping.EntityHandler;
-import jp.co.future.uroborosql.store.SqlResourceManager;
 import jp.co.future.uroborosql.store.SqlResourceManagerImpl;
+import jp.co.future.uroborosql.store.SqlResourceManager;
 
 /**
  * UroboroSQLを利用する際、初めに利用するクラス.
