+++ src/main/java/jp/co/future/uroborosql/store/SqlResourceManager.java
@@ -49,33 +49,6 @@ public interface SqlResourceManager {
 	List<String> getSqlPathList();
 
 	/**
-	 * SqlLoader の取得<br>
-	 * @return SqlLoader
-	 */
-	SqlLoader getSqlLoader();
-
-	/**
-	 * SqlLoader の設定<br>
-	 *
-	 * @param sqlLoader SqlLoader
-	 */
-	void setSqlLoader(SqlLoader sqlLoader);
-
-	/**
-	 * 起動時にSQLファイルをキャッシュするかどうかを取得する
-	 *
-	 * @return SQLをキャッシュする場合<code>true</code>
-	 */
-	boolean isCache();
-
-	/**
-	 * 起動時にSQLファイルをキャッシュするかどうか<BR>
-	 *
-	 * @param cache SQLをキャッシュする場合<code>true</code>
-	 */
-	void setCache(boolean cache);
-
-	/**
 	 * Dialectの取得
 	 *
 	 * @return Dialect
